Extract LSP runtime configuration into perl-lsp-config
