+++ Cargo.toml
@@ -52,7 +52,7 @@ members = [
     "crates/perl-lsp-input-validation",
     "crates/perl-lsp-limits",
     "crates/perl-lsp-uri",
-    "crates/perl-lsp-feature-policy",
+    "crates/perl-lsp-config",
     "crates/perl-diagnostics-codes",
     "crates/perl-position-tracking",
     "crates/perl-symbol-types",
@@ -205,7 +205,7 @@ allow = [
     "perl-lsp-launcher",
     "perl-lsp-limits",
     "perl-lsp-uri",
-    "perl-lsp-protocol",
+    "perl-lsp-config",
     "perl-lsp-transport",
 
     # Tier 6 — Module resolution higher tiers
@@ -313,7 +313,7 @@ perl-lsp-feature-contracts = { path = "crates/perl-lsp-feature-contracts", versi
 perl-lsp-cancellation = { path = "crates/perl-lsp-cancellation", version = "0.10.0" }
 perl-lsp-limits = { path = "crates/perl-lsp-limits", version = "0.10.0" }
 perl-lsp-uri = { path = "crates/perl-lsp-uri", version = "0.10.0" }
-perl-lsp-feature-grid = { path = "crates/perl-lsp-feature-grid", version = "0.10.0" }
+perl-lsp-config = { path = "crates/perl-lsp-config", version = "0.10.0" }
 perl-lsp-feature-profile = { path = "crates/perl-lsp-feature-profile", version = "0.10.0" }
 perl-lsp-feature-policy = { path = "crates/perl-lsp-feature-policy", version = "0.10.0" }
 perl-lsp-feature-governance = { path = "crates/perl-lsp-feature-governance", version = "0.10.0" }
